requirements.txt
requests==2.32.3
python-dotenv==1.0.1
Flask==2.3.2
Flask-Cors==4.0.0
certifi==2024.12.14
